Make tests depending on static resources work under IntelliJ and ant.
build.xml copies files that are not *.java in tests/com/splunk to build/tests/com/splunk.
IntelliJ knows about .csv, .json, and splunk.license as resource files to copy when it is building.
Fixed paths in ResultsReaderTest.java to use this.
Made LicenseTest use resource fetching instead of opening a file.