Made IndexTest.testUpload work with remote Splunk instances on different operating systems.
Before it used Java's file separator to join paths, which doesn't work if the SDK is running on Windows
and splunkd on Unix, or vice versa. Now it guesses the right separator from $SPLUNK_HOME, and
constructs the path accordingly, nor does it test for the file existing locally anymore.

Also added a note to the README about ant dist and ant build.