Make Service.search return an asynchronous search job instead of the results of a synchronous job.
Removed the three argument version. Renamed inputArgs to args in the two argument version and had it return a job.
Updated the docstrings and the changelog.
Updated the simple_search example to use the oneshot method instead.
Updated ServiceTest.testSearch to exercise it (and replaced its brittle logic with a call to a ResultsReader).