Fixed handling of segmentation for asynchronous jobs.
segmentation=none by default had mistakenly been put on JobCollection.create. It needs to go on Job.getEvents, job.getResults, and Job.getResultsPreview,
where it now is. The tests have been updated to fix this. Also, the tests were passing before because they did not wait for the asynchronous job to finish before
checking what the results contained, so it was always testing an empty stream.